Switch to `can-read?` not `can-write?`
original requestor intends to use a read-only template to be
copied. Thus no one can write to it (its a template, keep it clean) so
no one would be able to copy it, defeating the purpose.

Further, the ability to copy cards is based on `can-read?` so it wasn't
a meaningful or consistent protection:

```clojure
;; POST api/card/<id>/copy
;; based on `read-check`, not `write-check`
(api/defendpoint POST "/:id/copy"
  "Copy a `Card`, with the new name 'Copy of _name_'"
  [id]
  {id (s/maybe su/IntGreaterThanZero)}
  (let [orig-card (api/read-check Card id)
        new-name  (str (trs "Copy of ") (:name orig-card))
        new-card  (assoc orig-card :name new-name)]
    (create-card! new-card)))
```