Delay firing `:card-created` event until after tx
repeatedly calling `create-card!` fires the card created
event. Principally what this does is create the first revision of the
new card. But this work cannot find the newly created card until the
`db/transaction` is complete. And this transaction does not complete
until the new dashboard is fully created with all cards.

So just keep a copy of the new cards and fire the event after the tx
completes