Serdes v2: Inline TimelineEvents under Timelines (#26489)
This also changes the semantics of
`metabase.models.serialization.base/load-find-local` to return the whole
entity, not just the ID. `load-one!` always needs the whole entity (if
it exists) anyway, and some of the `load-find-local` implementations had
the entity already loaded but just passed along the ID. This saves some
double lookups, especially for inlining eg. `DashboardCard`s under
`Dashboard`.