Retrofit the release workflow to be useful for doing a pre-release (#26597)
* Retrofit the release workflow to be useful for doing a pre-release

* Distinguish between `metabase` main repo and the custom ones

* Truncate commit id to specific hash length

We need to make sure all images are tagged with exactly the same hash length
in order to avoid having multiple images for the same commit.

* Enforce full commit SHA-1 hash is used for an input

`actions/checkout@v3` will fail if full SHA-1 hash is not provided.
We want to use this implicit mechanism as an enforcing function
for uniform manual inputs.

* Never run script from the main repo

- This greatly simplifies things.
- The script is meant to be run from forks of Metabase for testing purposes and/or for custom builds.

* Safeguard against missing `CUSTOM_RELEASE_REPO`

Co-authored-by: Nemanja <31325167+nemanjaglumac@users.noreply.github.com>