Fix bug where grouping strategy would NPE when fingerprint was `nil`
The axes grouping relies on fingerprint data existing, AND for now only works on :type/Number columns, so this bugfix
guards against those situations where the logic would fail.

The bug that brought this to my attention came up when there was a Custom column. This set the col-meta count to be
one greater than the count of the 'real' columns, which threw off my old assumptions, where if the grouped-columns
count was less than the total col-meta count, you had a potential grouping. This broke in part because some custom
columns are essentially derived from a real column, and so will most likely be groupable with the column they're
derived from.

Anyway, this filters out columns that won't work, so we hopefully eliminate the problem.