Begin the proper png namespace for png downloads
This isn't actually working yet, just have the headers, need to study and figure out how to properly make the png
download use a streaming response from the query processor. This is important because we don't want to lock things up
downloading an image that happens to run a massive query.

So, the implementation that does work on this branch so far should not be considered shippable, just a prototype.