Solving Corrupt Cards #15222 caused by bad entries in json columns
This PR is the starting point to address 'corrupt cards'. This is kind of one bug, as the root cause is invalid json
in the app-db, causing the frontend to fail (in the case of no query-dataset being available), and the backend to
fail (in the case of un-normalizeable visualization-settings).

To reproduce, in a Clojure dev environment:

1. create and save a card and remember its id.
2. in the REPL, `(toucan.db/update! Card card-id :dataset_query "")`
 - note that if you do `(:dataset_query (toucan.db/select-one Card :id card-id))` you should see `nil`. This is
 because `cheshire.core/parse-string` will return `nil` for blank strings.
 - also note that you cannot update the :dataset_query key with `nil` directly, as there is a not-null constraint. So,
 the blank string basically circumvents this constraint, and is the root of our issue, I believe.
3. If you reload the question in your browser, you will see 'something's gone wrong'. You can further try to reproduce
based on the discussion in #15222, but this is a quick way to see this failure.

1. create and save a card and remember its id.
2. in the REPL, `(toucan.db/update! Card card-id :visualization_settings (cheshire.core/parse-string "{\"column_settings\":\"{\\\"(keyword [\\\\\\\"name\\\\\\\" \\\\\\\"ID\\\\\\\"])\\\":{\\\"number_separators\\\":\\\".\\\"}}\"}\n" keyword))`
3. Reload the question in the browser. You should see 'something went wrong' again.

I have created a post-select in the card api that will replace the :dataset_query with a 'dummy query' in the
situation where :dataset_query has a `nil` value. This will not recover the card's query, and I'm pretty sure the card
is effectively lost, but it should not break the frontend anymore, as the question can load properly, from the app's
perspective. _I do not know if there's a way to recover the query_. I am very open to suggestions or improvements on
this approach.

I have wrapped the visualization_settings normalization in a try catch that will log the failure but return the empty
map `{}`. This allows the app to load, but the visualization settings will be lost. Upon opening and saving the
question, new, valid viz settings will be written to the app-db. The original viz-settings will be lost, but the card
is still safe otherwise, and viz-settings are (somewhat) easier to recreate.