Disable field value scanning for inactive fields
Fixes #26863

Adds `last-used-at` to `metabase_fieldvalues` table.

This is used to track when full FieldValues are fetched,
which should always happen through `get-or-create-full-field-values!`
If an inactive FieldValues is queried, they will be updated immediately
 to be returned to the user and a new last-used-at will be set so that
 sync will start picking it up.

During field values sync, if `last-used-at` occurred more than 14 days
ago, then we will no longer query and synchronize this field until it is
queried for again.