Fix flaky `remap-human-readable-string-column-test`
Quite annoying test that would flake quite a bit in CI.

the problem is that the mechanism for `mt/with-column-remappings` wasn't
aware of the case when there were pre-existing mappings. It just blindly
added a new mapping. But when the middleware fetching the remappings, it
found the preexisting ones.

For the venue.name field in this test, that was ("20th Century
Cafe","25°","33 Taps","800 Degrees Neapolitan Pizzeria", ...). We were
passing in fake names like "apple", "banana", "kiwi" and it found no
remappings for those in the db.

So the fix is to check for pre-existing ones, and if found, use
`tu/with-temp-vals-in-db` to temporarily set them to the desired values,
otherwise use some new ones with `tt/with-temp*` as it was doing
previously.