[metabase-lib] Port number formatting to CLJC; use it from JS
This adds a cross-platform number formatting library in CLJC, which has
(nearly) identical output in both JVM and JS environments.

The only known differences are:
- JVM Clojure has `BigDecimal` support, so isn't limited to a `double`'s
  range.
- Exact spelling/capitalization of the full names of currencies differs.
    - JVM: "7.34 US Dollar"
    - JS: "7.34 US dollars"

Some future work is required to fully land this. In particular, the
`currency_in_header` logic should be moved to the table formatters, not
implemented on the JS side of this change.