Fix flaky `sync-recreate-field-values-test`
The tests assert that FieldValues exist for venues.price. But other
tests might delete these or sync hasn't finished. So ensure the sync
step has run.

This makes some of the tests kinda duplicative. But we'll live with
that. Example: the first test asserts values exist, deletes them,
asserts they are gone, syncs, and asserts they are back. The setup is
essentially tested so its not really a setup. But whatever.