Fixes to static-viz datetime formatter, in particular, issues #27020 and #27105  (#27339)
* Handle abbreviation and date-separator in a fn, where defaults are passed in in any place where nil COULD exist

Moving abbreviation and date-separator string changes into a post-processing fn allows each unit case to handle the
date-style default formatting, and eliminates the problem of date-style being `nil` in the let statement. It was this
scenario where `abbreviate` was `true`, but `date-style` was `nil`.

In many cases, if `date-style` is nil, then so is `abbreviate` and `date-separator`, in fact, the default from the
frontend sends nil for all of these keys. In such a case, none of the branches run and nil is safely passed
through. It is then handled appropriately with `(or ...)` expressions in each entry for the case statement.

But, it is indeed possible to have a nil `date-style` with non-nil `abbreviate` or `date-separator` keys. Now, the
post-process-date-style function is only ever called AFTER the default date-style is set in whatever case branch is needed.

* Turn frontend day format 'dddd' into backend format 'EEEE'

The frontend sends a datetime format string to the backend and some elements are not equivalent to the expectations
the java datetime formatter has.

* Unskip e2e reproductions

* Add a test for all 'written date' combos

* These tests fail b/c I fixed the 'default' from Abbreviated to non-abbreviated

* Add test 'matrix' for numerically formatted dates, and each separator

* Finally, add tests to confirm that custom-formatting is applied