allow for 3rd party logger configs (#27541)
```
docker run \
       -p 3000:3000 \
       -v $PWD/my_log4j2.xml:/tmp/my_log4j2.xml \
       -e JAVA_OPTS=-Dlog4j.configurationFile=/tmp/my_log4j2.xml \
       metabase/metabase:v0.45.1
```

From https://github.com/metabase/metabase/issues/27497

Note that this specifies `log4j.configurationFile` and not
"log4j2.configurationFile" that we set in our bootstrap.clj.

I'm not sure if this is supposed to work or not but we'll give it a
shot.

the Automatic Configuration section of
https://logging.apache.org/log4j/2.x/manual/configuration.html specifies
that:
1. Log4j will inspect the "log4j2.configurationFile" system property
and, if set, will attempt to load the configuration using the
ConfigurationFactory that matches the file extension. Note that this is
not restricted to a location on the local file system and may contain a
URL.

I'm not sure if that other property will work due to some other
mechanism or log4j2 has an undocumented back-compat mechanism.