Fix e2e Repro which works in 46 with new H2 driver, not 45 (old H2)
This e2e test failed in the 45 release branch because 46 (in progress on mater at time of writing) has an upgraded H2
driver, which was not backported.

This meant that the e2e test's expection of a graph.dimension of `EXTRACT(YEAR FROM CURRENT_TIMESTAMP)` would cause
the static-viz render to fail because the actual dimension would be `YEAR(CURRENT_TIMESTAMP())`.

The fix to the underlying problem is still valid, and this reproduction, with the correct dimension, tests the fix and
behaves correctly.