Do full syncs on both "test-data" and "sample-dataset"
previously it was just on "test-data". So tests run against a fresh
sample-dataset might only have a schema sync. And this makes tests like
`metabase.query-processor.util.add-alias-info-test/join-source-query-join-test`
fail because the expect

```
{:base_type         :type/Text
 :coercion_strategy nil
 :database_type     "CHARACTER VARYING"
 :effective_type    :type/Text
 :name              "CATEGORY"
 :semantic_type     :type/Category}
```
to be fully formed. But semantic_type is only set by full analysis in
sync so a quick sync leaves that out.

Running more tests (presumably a sync test, or likely a test that just
happened to call sync) ran before this one.