Handle race condition in bigquery cancellation
This was a big pain. We have some uncoordinated threads going on. The
long and short is that with the changes to run fewer tests in CI, we
were getting consistent failures in bigquery's `query-cancel-test`.

> ERROR in metabase.driver.bigquery-cloud-sdk-test/query-cancel-test
>                                       (bigquery_cloud_sdk.clj:235)
> Uncaught exception, not in assertion.
>
>                clojure.lang.ExceptionInfo: Error executing query: null

This would also happen on the _FIRST_ run of all bigquery tests in the
repl

```
bigquery-cloud-sdk-test=> (run-tests)

Testing metabase.driver.bigquery-cloud-sdk-test
[BigQuery] CREATE VIEW `v3_test_data.view_DWPNQKUJDDUEACNYKGML` AS SELECT v.id AS id, v.name AS venue_name, c.name AS category_name FROM `metabase-bigquery-driver.v3_test_data.venues` v LEFT JOIN `metabase-bigquery-driver.v3_test_data.categories` c ON v.category_id = c.id ORDER BY v.id ASC LIMIT 3
[BigQuery] DROP VIEW IF EXISTS `v3_test_data.view_DWPNQKUJDDUEACNYKGML`

ERROR in (query-cancel-test) (bigquery_cloud_sdk.clj:235)
Uncaught exception, not in assertion.
expected: nil
  actual: clojure.lang.ExceptionInfo: Error executing query: null
{:type :invalid-query, :sql "-- Metabase\nSELECT `v3_sample_dataset.orders`.`id` AS `id`, `v3_sample_dataset.orders`.`user_id` AS `user_id`, `v3_sample_dataset.orders`.`product_id` AS `product_id`, `v3_sample_dataset.orders`.`subtotal` AS `subtotal`, `v3_sample_dataset.orders`.`tax` AS `tax`, `v3_sample_dataset.orders`.`total` AS `total`, `v3_sample_dataset.orders`.`discount` AS `discount`, `v3_sample_dataset.orders`.`created_at` AS `created_at`, `v3_sample_dataset.orders`.`quantity` AS `quantity` FROM `v3_sample_dataset.orders` LIMIT 1048575", :parameters nil}
 at metabase.driver.bigquery_cloud_sdk$throw_invalid_query.invokeStatic (bigquery_cloud_sdk.clj:235)
    metabase.driver.bigquery_cloud_sdk$throw_invalid_query.invoke (bigquery_cloud_sdk.clj:234)
    metabase.driver.bigquery_cloud_sdk$execute_bigquery.invokeStatic (bigquery_cloud_sdk.clj:277)
    ...
Caused by: java.util.concurrent.CancellationException: null
 at java.util.concurrent.FutureTask.report (FutureTask.java:121)
    java.util.concurrent.FutureTask.get (FutureTask.java:191)
    ...

Ran 19 tests containing 93 assertions.
0 failures, 1 errors.
{:test 19, :pass 92, :fail 0, :error 1, :type :summary}
```

Subsequent runs of `(query-cancel-test)` would pass.

This is because we have a few threads doing work:
- one thread is doing the query work:
`(future (.query client (.build request) (u/varargs BigQuery$JobOption)))`
- one thread is watching for cancelled-chan for a cancellation and then
attempting to cancel the thread above.

The error seen here is `Caused by:
java.util.concurrent.CancellationException: null` which happens when
derefing a future that was cancelled with no value.

I'm a bit wishy-washy on what exactly is going on here. So a few notes:
- I'm not sure why the future which watches the cancel-chan derefs the
`res-future`. No one is watching the future and no one cares about its
return value
- But the main body deref's the `@res-fut` which returns a
`TableResult`. But if the future is cancelled before anything comes
back (due to the race) there's kinda nothing to do here.

I had done three different attempts to solve this:
- make a protocol that delegates to the TableResult and when we cancel
the future without a TableResult can pass an empty result to indicate no
results
- after making this i saw that there's an EmptyTableResult we could add
in.

Both of these suffer from a lack of metadata. And in the world of
recording metadata and especially models, this could be a way to blank
out all of the columns. So this isn't a great way to go.

But fundamentally we want cancelled queries to throw. If a query is
cancelled we can't reasonably continue through the qp pipeline. So its
got to revert control higher up. So this third attempt is a meaningful
and correct solution, not just "catch the error" and continue.

A _valid_ way to handle a cancelled query is to throw with a message.

```
 :context :ad-hoc,
 :error "ERROR: Query (10601832) cancelled on user's request",
 :row_count 0,
```