Create migration to add activity_log
Differences between activity_log (new) and activity (old):

- drop `custom_id` from activity, it's unused
- allow `details` to be null
- add nullable `end_timestamp` which can be used for long running events (eg. a slow query)

`details` is a text column in the same way `visualization_settings` is a text column -> it will be storing a json
blob, which will be handled by pre-insert and post-selects, I think.

What I'm not sure about still:
- can we use proper json columns now that we have newer H2?
- should end_timestamp actually exist? We could model the same kind of idea by instead creating 2 events:
  - :query_start (timestamp then represents the instant a query began)
  - :query_end (timestamp of a different event represents the end)
  - this would let us also create events like :query_failed
- if we have an open json blob in 'details', we might discover that there are certain keys that are essentially
expected -> can we anticipate some of those and make them columns on the table directly?