Hook up card params endpoint for current style as well as new
Previously the new `"/:card-id/params/:param-key/values"` endpoint
handled the "static-list" and the "card" new styles of parameters. But
it did not handle the "current" style to look up field values. The FE
was expected to hit `api/field/:id/values`.

This change makes the card endpoint work more like the dashboard
endpoint which did handle all three. It cannot be exactly the same
though as we aren't looking on dashboard's for params and we don't have
dashcards.

So we have to skip a lot of the chain filter stuff but we end up using
the same stuff at the end.