Fix failing bigquery test (#27837)
Bigquery was checking that we can set a project-id in the db details of
a bigquery database. Normally this comes from the json token:

```
export MB_BIGQUERY_CLOUD_SDK_TEST_SERVICE_ACCOUNT_JSON='{
  "type": "service_account",
  "project_id": "metabase-bigquery-driver", // <- ordinarily source of truth
  "private_key_id": "keykeykeykeykeykeykeykey",
  "private_key": "-----BEGIN PRIVATE KEY----- ...."
```

The db details normally connects to project-id
"metabase-bigquery-driver" and this test swaps out to
"bigquery-public-data" with a schema filter on chicago_taxi_trips.

to prove it works it selects from that dataset:

```
(mt/first-row
  (mt/run-mbql-query taxi_trips
    {:filter [:= [:field (mt/id :taxi_trips :unique_key) nil]
                 "67794e631648a002f88d4b7f3ab0bcb6a9ed306a"]}))
```

The test fails because this row seems to no longer be in the
results. The query isn't failing, it's just returning no rows (and
presumably because the row with this id from 2014 is no longer in the
dataset.)

I ran the same with with a limit 4, see it's working and grab one of
those rows and use the id as in the filter clause.

old: "67794e631648a002f88d4b7f3ab0bcb6a9ed306a" ("2014-01-12T00:45:00Z")
new: "ff0b96c0cada768361b1c9341e11905254644afb" ("2016-09-20T07:15:00Z")