only defsetting changes
tracking down why https://github.com/metabase/metabase/pull/27857 is
failing in CI. Seems to be an issue with settings for ee out of
sync. Want to reduce down to just the changes to defsetting to see if we
can isolate if it is some of the setting changes or just this
mechanism.