Group all axes in one shot, and make overlap take the row vals
Work in progress implementation here. If it's too much trouble, it might be best to default everything to either use
the explicitly set position (if user sets Right or Left), and otherwise jsut use :left, so that we at least have
sensible and usable renders, with relatively low effort.

If this works, it'll be nice though!