[metabase-lib] Overhaul infra; port part of `Question.ts`
This overhauls the approach to integrating CLJC logic into the
metabase-lib classes.

Rather than the `cljs-bean` powered approach, this new
infrastructure actually converts from JS data to CLJS data. The
TS classes hold an opaque CLJS map as their `_inner` data.
There's no more automagic conversion on `de/defn` - it's a wrapper
for `mu/defn` now.

Many of the functions are getters and setters for nested fields; the
new `de/def-getters-and-setters` generates those from a list of
names and `get-in` paths.

Some of `Question.ts` is not portable now because it's tied in with
`Query.ts`. These two files will probably each move in pieces as the
other unblocks them.