Ensure query action's db matches model id
This is only a half-way measure. Needs some FE loving as well.

The +New -> Action workflow lets you create an action. This starts with
the user choosing the database, writing the action. On save it prompts
you to choose a model it should live under. But this model's db and the
db chosen originally are not checked in the FE.

Ideally we only show models on that existing db, but it kinda seems
weird to have an action do things to a random table and it live under a
model that just needs to be on the db.