Optimize formatNumber and formatDate*
These were causing perceptible lag on dashboards. The culprit is
not the formatting itself, but conversion of the `options`
objects into Clojure-friendly form.

In particular, transforming `"snake_case"` and `"camelCase"` strings
into `:kebab-case` requires a lot of string munging. These are fairly
hot functions, called on each cell of a table every time they re-render,
so it pays to be fast here.

This improves the time spent here by 2.5-3x on my machine, and that's
likely to scale up even more with larger tables.